requirements.txt
bleach                  ==2.1.3
certifi                 ==2017.7.27.1
chardet                 ==3.0.4
coreapi                 ==2.3.1
coreschema              ==0.0.4
crcmod                  ==1.7
defusedxml              ==0.5.0
diff-match-patch        ==20121119
Django                  ==1.10.8
django-cors-headers     ==2.1.0
django-crispy-forms     ==1.6.1
django-filter           ==1.0.4
django-formtools        ==2.0
django-guardian         ==1.4.9
django-haystack         ==2.8.1
django-import-export    ==1.0.0
django-qiniu-storage    ==2.3.1
django-redis            ==4.8.0
django-reversion        ==2.0.9
djangorestframework     ==3.6.3
djangorestframework-jwt ==1.11.0
drf-extensions          ==0.3.1
et-xmlfile              ==1.0.1
html5lib                ==1.0.1
idna                    ==2.6
itypes                  ==1.1.0
jdcal                   ==1.3
Jinja2                  ==2.9.6
MarkupSafe              ==1.0
odfpy                   ==1.3.6
openpyxl                ==2.5.0
pip                     ==19.0.3
PyJWT                   ==1.5.3
python-dateutil         ==2.6.1
PyYAML                  ==3.12
qiniu                   ==7.2.0
redis                   ==2.10.6
requests                ==2.18.4
setuptools              ==40.8.0
six                     ==1.11.0
tablib                  ==0.12.1
unicodecsv              ==0.14.1
uritemplate             ==3.0.0
urllib3                 ==1.22
virtualenv              ==16.4.3
virtualenvwrapper-win   ==1.2.5
webencodings            ==0.5.1
xlrd                    ==1.1.0
xlwt                    ==1.2.0
